grub_old.cfg: make provisioning IO use ttyS0
The provisioning tool is not configured to use any particular TTY device
for its console. In QEMU (at least), non-graphical IO is configured by
default to use ttyS0, but the provisioning scripts output over
/dev/console, which is set to something else like tty0.

As a result, users cannot interactively provision using QEMU without
manually restarting the provisioning script.

Pass the `console=ttyS0` kernel argument to the provisioning kernel, to
align the tool to ttyS0. This is exactly the same argument as we pass to
the non-provisioning kernel.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>