initrdscripts: allow the user to select provisioning block device
There is currently no way for a user to express that they would like the
provisioning script to install NILRT to a non-standard block device,
without setting back-end installation variables.

For interactive cases, collect and present the user with a list of all
valid installation block devices on the system. Allow the user to
declare which device should be provisioned.

The TARGET_DISK environment variable can bypass this selection for
non-interactive installation cases.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>