initrdscripts: restructure init-restore-mode
The init-restore-mode recipe dumps all its component files in the root
directory; the result is kind of a mess.

Clean up the package layout by distributing files into more standard
locations. The top-level provisioning scripts are put in /sbin. The
source scripts and default files are placed in namespaced locations in
/lib and /usr/share respectively.

Also reconcile the package files to account for the restructuring.

Nothing in this commit is intended to change behavior.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>