nilrt-core-image: set core FSTYPE to tar.gz
tar.bz2 FSTYPES started getting thrown around in image recipes to
facilitate the RAUC bundle implementation, and they infected their way
into the refactored hardknott images. tar.gz is a little easier to work
with and other parts of the hardknott pipeline (the CDF packing logic in
nilrt-base-system-image.bb) requires that the runmode rootfs be tar.gz
anyway.

So prefer to make tar.gz the core FSTYPE. RAUC-enabled recipes can add
other rootfs types in an includes, when we get to that work.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>