glibc: restore locale.alias file for subpackage
The stash_locale_package_cleanup() function from OE-core's
glibc-package.inc removes the /usr/share/locale directory from the
packages/ staging dir prior to packaging. So when locale-alias tries to
capture the locale.alias file, it finds that the alias file is gone.

The alias files has been copied to the "locale stash" directory in the
workspace. So we can restore it back into the packaging stage prior to
package-split.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>