packagegroup-ni-runmode: remove libfmi-dev dep
The runmode packagegroup has an RDEPENDS on libfmi-dev, meaning that it
will be installd to the runmode image by default.

There is no reason to do this. If Veristand requires libfmi-dev, then
the Veristand packages should express an IPK dependency on it, and it
will be installed to the runmode image at runtime when Veristand is
installed.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>