apr: remove bbappend
The meta-nilrt apr bbappend adds additional CONFIGUREVARS to the apr
do_configure, which supposedly enable pthread support in apr - which was
needed to resolve a runtime hang in apache2.

As of OE kirkstone and apr 1.7, these additional configs break apr
do_compile. Further, the configs were added many years ago, and it is
unclear if they are still needed for apache.

Remove the configs. If apache2 is still broken, we'll need to fix it
using a different approach anyway.

Remove the entire bbappend, since it is now empty.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>