tbtadm: fix do_configure pkgconfig error
tbtadm's CMakeLists tries to set a variable PKG_CONFIG_UDEV_DIR by
querying the 'udevdir' variable from the udev.pc pkgconfig definition
file. In our OE build, that is provided by eudev.

However, the tbtadm recipe doesn't use the pkgconfig.bbclass. So though
the .pc file is installed to the sysroot, it is not linked to cmake. So
the variable is unpopulated, causing other CMakeLists files in the
tbtadm source which use the variable to fail with cryptic errors about
"install PROGRAMS given no DESTINATION". (The DESTINATION should have
been the PKG_CONFIG_UDEV_DIR.

Give tbtadm the pkgconfig.bbclass to resolve this issue.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>