salt: fix append-operator use
salt-common and salt-minion try to append additional python packages to
the salt RDEPENDS using both the `:append` override, and the
append-operator (+=). Not only does OE kirkstone complain about this
usage, but it is actually wrong. Using both misapplies the packages to
an empty RDEPENDS subvariable.

Drop the :append override, and instead append directly to the salt
subpackage's RDEPENDS variable.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>